pyproject.toml
[project]
name = "llmproject"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "jsons>=1.6.3",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "telebot>=0.0.5",
]
